Handle multiple clients and check method support
This finally correctly asks each client individually depending on if the
method is supported or not. Before it was a primitive approach where
only the first client got checked to support the method.

This might be a little slower than the `vim.lsp.buf_request_sync`
method. I'm not 100% sure about its exact implementation and if it asks
the clients in parallel or iterative. But as a buffer usually never has
many clients attached, this should not be a big factor. Even if all of
them support all provider methods.